Moved old template system to PagesCore::Deprecated
