+++ lib/pages_core/deprecated/templates.rb
@@ -1,0 +1,14 @@
+# encoding: utf-8
+
+load "pages_core/deprecated/templates/block_configuration.rb"
+load "pages_core/deprecated/templates/configuration_proxy.rb"
+load "pages_core/deprecated/templates/configuration_handler.rb"
+load "pages_core/deprecated/templates/configuration.rb"
+load "pages_core/deprecated/templates/template_configuration.rb"
+
+module PagesCore
+  module Deprecated
+    module Templates
+    end
+  end
+end
