+++ lib/pages_core/deprecated/templates/block_configuration.rb
@@ -1,0 +1,20 @@
+# encoding: utf-8
+
+module PagesCore
+  module Deprecated
+    module Templates
+      # Configuration for the blocks on an individual template
+      class BlockConfiguration
+        attr_reader :name, :title, :description, :optional, :enforced
+
+        def small?
+          @size == :small
+        end
+
+        def large?
+          small? ? false : true
+        end
+      end
+    end
+  end
+end
